+++ Terra/src/Terra/Events/MouseEvent.h
@@ -7,11 +7,11 @@ namespace Terra {
 	class MouseMovedEvent : public Event
 	{
 	public:
-		MouseMovedEvent(float x, float y)
+		MouseMovedEvent(int x, int y)
 			: m_MouseX(x), m_MouseY(y) {}
 
-		inline float GetX() const { return m_MouseX; }
-		inline float GetY() const { return m_MouseY; }
+		inline int GetX() const { return m_MouseX; }
+		inline int GetY() const { return m_MouseY; }
 
 		std::string ToString() const override
 		{
@@ -24,13 +24,13 @@ namespace Terra {
 		EVENT_CLASS_CATEGORY(EventCategoryMouse | EventCategoryInput)
 
 	private:
-		float m_MouseX, m_MouseY;
+		int m_MouseX, m_MouseY;
 	};
 
 	class RawMouseMovedEvent : public Event
 	{
 	public:
-		RawMouseMovedEvent(float xOffset, float yOffset)
+		RawMouseMovedEvent(int xOffset, int yOffset)
 			: m_MouseXOffset(xOffset), m_MouseYOffset(yOffset) {}
 
 		inline int GetX() const { return m_MouseXOffset; }
@@ -53,11 +53,11 @@ namespace Terra {
 	class MouseScrolledEvent : public Event
 	{
 	public:
-		MouseScrolledEvent(float xOffset, float yOffset)
+		MouseScrolledEvent(int xOffset, int yOffset)
 			: m_XOffset(xOffset), m_YOffset(yOffset) {}
 
-		inline float GetXOffset() const { return m_XOffset; }
-		inline float GetYOffset() const { return m_YOffset; }
+		inline int GetXOffset() const { return m_XOffset; }
+		inline int GetYOffset() const { return m_YOffset; }
 
 		std::string ToString() const override
 		{
@@ -70,7 +70,7 @@ namespace Terra {
 		EVENT_CLASS_CATEGORY(EventCategoryMouse | EventCategoryInput)
 
 	private:
-		float m_XOffset, m_YOffset;
+		int m_XOffset, m_YOffset;
 	};
